[Remove ViewUrl, 예외 처리 적용]
- 불필요하게 적용되었던 ViewUrl이 제거되었습니다.
  - 그에 따른 레포지토리 및 테스트 코드도 제거되었습니다.
  - Url에 viewCount 필드를 추가하고, 이를 활용하게 되었습니다.

- 특정 상황에 맞는 예외가 추가되었습니다.
  - Url 추가 시 중복이면 발생하는 DuplicateUrlException
  - Url 검색 시 없으면 발생하는 NoSuchUrlException

- 추가된 예외를 활용하여 @RestControllerAdvice를 통한 전역 예외 핸들러가 추가되었습니다.
  - 이를 통해 서비스에서 컨트롤러로 주던 에러 상황에서의 특정 값이 사라졌습니다. 단지 컨트롤러로 추가된 예외를 던지게 됩니다.

+++ src/main/java/com/toy/shorturl/dto/ViewcountResponse.java
@@ -2,4 +2,4 @@ package com.toy.shorturl.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record ViewcountResponse(@NotBlank String encodedUrl, int viewCount) {}
+public record ViewcountResponse(@NotBlank String encodedUrl, long viewCount) {}
